Match ports to USB Identifier for OpenFIRE firmware
Code checks if Gun is running OpenFIRE firmware. Then if it is, it checks the the TinyUSB identifier numbers to see if they are different. If they are it assigns a port to match assigned ID's. If not, prints an error telling the user to set TinyUSB identifiers in OpenFIRE App.